Refactor loader into modular package
